Rename UserProfileCommon to UserProfileFormFields
